+++ src/adv_file.h
@@ -57,13 +57,13 @@ namespace AdvLib
 			int AddFileTag(const char* tagName, const char* tagValue);
 			int AddUserTag(const char* tagName, const char* tagValue);
 			
-			void BeginFrame(long long timeStamp, unsigned int elapsedTime, unsigned int exposure);
+			void BeginFrame(__int64 timeStamp, unsigned int elapsedTime, unsigned int exposure);
 			void AddFrameStatusTag(unsigned int tagIndex, const char* tagValue);
 			void AddFrameStatusTagMessage(unsigned int tagIndex, const char* tagValue);
 			void AddFrameStatusTagUInt8(unsigned int tagIndex, unsigned char tagValue);
 			void AddFrameStatusTagUInt16(unsigned int tagIndex, unsigned short tagValue);			
 			void AddFrameStatusTagUInt32(unsigned int tagIndex, unsigned int tagValue);
-			void AddFrameStatusTagUInt64(unsigned int tagIndex, long long tagValue);
+			void AddFrameStatusTagUInt64(unsigned int tagIndex, __int64 tagValue);
 			void AddFrameStatusTagReal(unsigned int tagIndex, float tagValue);
 			void AddFrameImage(unsigned char layoutId, unsigned short* pixels, unsigned char pixelsBpp);
 			void EndFrame();
